Added custom error handling hook
